+++ engine/runtime-compiler/src/main/java/org/enso/compiler/pass/analyse/TailCall.java
@@ -18,6 +18,7 @@ import org.enso.compiler.core.ir.Warning;
 import org.enso.compiler.core.ir.expression.Application;
 import org.enso.compiler.core.ir.expression.Case;
 import org.enso.compiler.core.ir.expression.Comment;
+import org.enso.compiler.core.ir.expression.IfThenElse;
 import org.enso.compiler.core.ir.module.scope.Definition;
 import org.enso.compiler.core.ir.module.scope.definition.*;
 import org.enso.compiler.pass.IRPass;
@@ -206,10 +207,20 @@ public final class TailCall implements MiniPassFactory {
           // Note [Call Argument Tail Position]
           p.arguments().foreach(a -> markAsTail(a));
         }
+        case IfThenElse ite -> {
+          if (isInTailPos) {
+            markAsTail(ite);
+            // Note [Analysing Branches in Conditional Expressions]
+            markAsTail(ite.trueBranch());
+            if (ite.falseBranchOrNull() != null) {
+              markAsTail(ite.falseBranchOrNull());
+            }
+          }
+        }
         case Case.Expr e -> {
           if (isInTailPos) {
             markAsTail(ir);
-            // Note [Analysing Branches in Case Expressions]
+            // Note [Analysing Branches in Conditional Expressions]
             e.branches().foreach(b -> markAsTail(b));
           }
         }
@@ -246,6 +257,15 @@ public final class TailCall implements MiniPassFactory {
         Expression expression, java.util.Map<IR, Boolean> tailCandidates) {
       switch (expression) {
         case Function function -> collectTailCandicateFunction(function, tailCandidates);
+        case IfThenElse ite -> {
+          if (isInTailPos) {
+            // Note [Analysing Branches in Conditional Expressions]
+            tailCandidates.put(ite.trueBranch(), true);
+            if (ite.falseBranchOrNull() != null) {
+              tailCandidates.put(ite.falseBranchOrNull(), true);
+            }
+          }
+        }
         case Case caseExpr -> collectTailCandidatesCase(caseExpr, tailCandidates);
         case Application app -> collectTailCandidatesApplication(app, tailCandidates);
         case Name name -> collectTailCandidatesName(name, tailCandidates);
@@ -345,7 +365,7 @@ public final class TailCall implements MiniPassFactory {
       switch (caseExpr) {
         case Case.Expr expr -> {
           if (isInTailPos) {
-            // Note [Analysing Branches in Case Expressions]
+            // Note [Analysing Branches in Conditional Expressions]
             expr.branches()
                 .foreach(
                     b -> {
@@ -358,7 +378,7 @@ public final class TailCall implements MiniPassFactory {
       }
     }
 
-    /* Note [Analysing Branches in Case Expressions]
+    /* Note [Analysing Branches in Conditional Expressions]
      * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      * When performing tail call analysis on a case expression it is very
      * important to recognise that the branches of a case expression should all