+++ CHANGELOG.md
@@ -174,6 +174,12 @@
 [11374]: https://github.com/enso-org/enso/pull/11374
 [11671]: https://github.com/enso-org/enso/pull/11671
 
+#### Enso Language & Runtime
+
+- [Unify internal handling of if/then/else and case statement][11365]
+
+[11365]: https://github.com/enso-org/enso/pull/11365
+
 # Enso 2024.4
 
 #### Enso IDE